requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.4.4
requests==2.31.0
stripe==5.4.0
